Quickhack 2021 - Calm Mind
#### Theme: Mental Health

#### Product Name: Calm Mind

**CALM MIND** is an interactive platform opened to all students with the aim of motivating students when times are hard as well as forming a new community based on the topic of the day. Through this platform, students are free to share their interests, ideas and current thoughts with other students. Our team hope that by providing a space for students to express their current thoughts and feelings while encourage engagement between them, it will bring down the number of adolescents suffering from mental health illnesses. 

Besides that, Word Cloud is a feature where students are able to enter their current feeling as a keyword. The more the keyword is being entered, the larger the keyword will display. A frequently asked question (FAQ) page is provided if new users have any common queries regarding **CALM MIND**. Our platform have a total of 6 pages which are the *home page*, forum page, *FAQ page*, *About Us page*, *login page* and a *chat discussion page* which can be found at the bottom right of the forum page. 



Features Highlight

- Everyday, a simple question will be posted
- everything can be answered anonymously
- the next day, a forum will be open and a hashtag feature will be introduced to post the relevant picture